Add checksum to transaction logs
Checksums are calculated over each transaction so we can validate the integrity of the stored files when doing recovery. CRC32C was chosen because it should theoretically give better performance when working in buffered reader/writer as we do.

Also, remove masterId and authorId from transactions since that is no longer the source for checksums.

+++ community/record-storage-engine/src/main/java/org/neo4j/internal/batchimport/AdditionalInitialIds.java
@@ -41,7 +41,7 @@ public interface AdditionalInitialIds
 {
     long lastCommittedTransactionId();
 
-    long lastCommittedTransactionChecksum();
+    int lastCommittedTransactionChecksum();
 
     long lastCommittedTransactionLogVersion();
 
@@ -59,7 +59,7 @@ public interface AdditionalInitialIds
         }
 
         @Override
-        public long lastCommittedTransactionChecksum()
+        public int lastCommittedTransactionChecksum()
         {
             return TransactionIdStore.BASE_TX_CHECKSUM;
         }
